thangnv62 update first code

+++ src/main/java/com/vietfintex/marketplace/persistence/repo/GroupClubCustomRepo.java
@@ -6,5 +6,5 @@ import java.math.BigInteger;
 import java.util.List;
 
 public interface GroupClubCustomRepo {
-    List<GroupClub> getListGroupClub(Long groupId, String privacy, String keyword,BigInteger ownerId);
+    List<GroupClub> getListGroupClub(Long groupId, String privacy, String keyword,Long ownerId);
 }
